# open-api/passtheo-content-service-openapi.yaml
openapi: 3.1.0
info:
  title: PassTheo Content Service
  description: |
    The core learning engine for PassTheo. Serves questions, tracks progress,
    runs mock exams, manages streaks, awards achievements, and generates study plans.
    
    Content hierarchy: Country → ProductType → Product → Domain → Topic → Question
    Language is Strapi i18n, orthogonal to hierarchy.
  version: 1.0.0
servers:
  - url: http://localhost:8087
    description: Local development
...
paths:
...
  /api/exams/mock/start:
    post:
      operationId: startMockExam
      tags: [Exams]
      summary: Start a full mock exam
      description: |
        Generates a complete mock exam with questions distributed across
        all domains according to the ExamConfig. No feedback during exam.
        Timer starts immediately.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExamRequest'
            example:
              productCode: "auto-b"
              locale: "en"
      responses:
        '200':
          description: Exam started with all questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Exam'
              example:
                success: true
                data:
                  examId: "880e8400-e29b-41d4-a716-446655440001"
                  totalQuestions: 50
                  timeLimitSeconds: 1800
                  passScore: 44
                  startedAt: "2026-03-19T10:00:00Z"
                  expiresAt: "2026-03-19T10:30:00Z"
                  questions:
                    - strapiQuestionId: "q-101"
                      questionText: "What does this sign mean?"
                      interactionType: "multiple_choice"
                      imageUrl: "https://cdn.passtheo.nl/questions/q101.png"
                      answerOptions:
                        - id: "a1"
                          text: "Give way"
                        - id: "a2"
                          text: "No entry"
                        - id: "a3"
                          text: "One way"
                      questionOrder: 1
                      domainCode: "verkeersborden"
                timestamp: "2026-03-19T10:00:00Z"
        '403':
          description: Premium required — free users cannot take mock exams
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
              example:
                type: "https://api.passtheo.nl/errors/premium-required"
                title: "Premium required"
                status: 403
                detail: "Upgrade to Pro to take mock exams."
  /api/exams/mock/{examId}/submit:
    post:
      operationId: submitMockExam
      tags: [Exams]
      summary: Submit all exam answers at once
      description: |
        Submits all answers, calculates score, determines pass/fail,
        updates progress, publishes exam.completed event.
      security:
        - bearerAuth: []
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitExamRequest'
      responses:
        '200':
          description: Exam result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExamResult'
              example:
                success: true
                data:
                  examId: "880e8400-e29b-41d4-a716-446655440001"
                  passed: true
                  correctCount: 46
                  totalQuestions: 50
                  passScore: 44
                  scorePercent: 92.0
                  timeTakenSeconds: 1523
                  domainBreakdown:
                    - domainCode: "verkeersborden"
                      domainName: "Verkeersborden"
                      correct: 8
                      total: 9
                      accuracyPercent: 88.9
                    - domainCode: "voorrang"
                      domainName: "Voorrang"
                      correct: 7
                      total: 8
                      accuracyPercent: 87.5
                  wrongAnswers:
                    - strapiQuestionId: "q-115"
                      questionText: "Wat is de maximale snelheid..."
                      yourAnswer: { selectedOptionId: "a2" }
                      correctAnswer: { selectedOptionId: "a3" }
                      explanation:
                        text: "Op autosnelwegen is de maximale..."
                  readinessUpdate:
                    previousScore: 68.5
                    newScore: 76.2
                    label: "ALMOST_READY"
                  newAchievements:
                    - code: "first_exam"
                      name: "Exam ready"
                timestamp: "2026-03-19T10:25:23Z"
  /api/exams/history:
    get:
      operationId: getExamHistory
      tags: [Exams]
      summary: List past mock exam attempts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
        - name: productCode
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Exam history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ExamHistoryList'

  # ─── PROGRESS ───
...
  /api/streaks/me:
    get:
      operationId: getMyStreak
      tags: [Streaks]
      summary: Get current streak status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/UserHeader'
        - name: productCode
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Streak status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Streak'
              example:
                success: true
                data:
                  currentStreak: 12
                  longestStreak: 18
                  totalStudyDays: 34
                  lastStudyDate: "2026-03-19"
                  freezeSlotsAvailable: 2
                  freezeSlotsUsed: 1
                  studiedToday: true
                  streakAtRisk: false
                  lastSevenDays: [true, true, false, true, true, true, true]
                timestamp: "2026-03-19T10:00:00Z"

  # ─── ACHIEVEMENTS ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    UserHeader:
      name: X-Keycloak-User-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    StartExamRequest:
      type: object
      required: [productCode]
      properties:
        productCode:
          type: string
        locale:
          type: string
          default: "nl"
    SubmitExamRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [strapiQuestionId, answer]
            properties:
              strapiQuestionId:
                type: string
              answer:
                type: object
              timeTakenMs:
                type: integer
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    ApiResponse_Exam:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/ExamDto' }
        timestamp: { type: string, format: date-time }
    ApiResponse_ExamResult:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/ExamResultDto' }
        timestamp: { type: string, format: date-time }
    ApiResponse_ExamHistoryList:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { $ref: '#/components/schemas/ExamHistorySummaryDto' } }
        timestamp: { type: string, format: date-time }
...
    ApiResponse_Streak:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/StreakDto' }
        timestamp: { type: string, format: date-time }
...
    QuestionDto:
      type: object
      properties:
        strapiQuestionId: { type: string }
        questionText: { type: string }
        interactionType: { type: string, enum: [multiple_choice, yes_no, fill_in_number, tap_on_image, drag_checkmark, drag_numbers] }
        imageUrl: { type: string, nullable: true }
        videoUrl: { type: string, nullable: true }
        answerOptions:
          type: array
          nullable: true
          items:
            type: object
            properties:
              id: { type: string }
              text: { type: string }
              imageUrl: { type: string, nullable: true }
        imageRegions:
          type: array
          nullable: true
          items:
            type: object
            properties:
              id: { type: string }
              xPercent: { type: number }
              yPercent: { type: number }
              widthPercent: { type: number }
              heightPercent: { type: number }
        dragTargets:
          type: array
          nullable: true
          items:
            type: object
            properties:
              id: { type: string }
              label: { type: string }
              imageUrl: { type: string, nullable: true }
        questionOrder: { type: integer }
...
    ExamDto:
      type: object
      properties:
        examId: { type: string, format: uuid }
        totalQuestions: { type: integer }
        timeLimitSeconds: { type: integer }
        passScore: { type: integer }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        questions:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/QuestionDto'
              - type: object
                properties:
                  domainCode: { type: string }
    ExamResultDto:
      type: object
      properties:
        examId: { type: string, format: uuid }
        passed: { type: boolean }
        correctCount: { type: integer }
        totalQuestions: { type: integer }
        passScore: { type: integer }
        scorePercent: { type: number }
        timeTakenSeconds: { type: integer }
        domainBreakdown:
          type: array
          items:
            type: object
            properties:
              domainCode: { type: string }
              domainName: { type: string }
              correct: { type: integer }
              total: { type: integer }
              accuracyPercent: { type: number }
        wrongAnswers:
          type: array
          items:
            type: object
            properties:
              strapiQuestionId: { type: string }
              questionText: { type: string }
              yourAnswer: { type: object }
              correctAnswer: { type: object }
              explanation: { type: object }
        readinessUpdate:
          type: object
          properties:
            previousScore: { type: number }
            newScore: { type: number }
            label: { type: string }
        newAchievements:
          type: array
          items:
            $ref: '#/components/schemas/EarnedAchievementDto'
    ExamHistorySummaryDto:
      type: object
      properties:
        examId: { type: string, format: uuid }
        passed: { type: boolean }
        correctCount: { type: integer }
        totalQuestions: { type: integer }
        scorePercent: { type: number }
        completedAt: { type: string, format: date-time }
...
    StreakDto:
      type: object
      properties:
        currentStreak: { type: integer }
        longestStreak: { type: integer }
        totalStudyDays: { type: integer }
        lastStudyDate: { type: string, format: date }
        freezeSlotsAvailable: { type: integer }
        freezeSlotsUsed: { type: integer }
        studiedToday: { type: boolean }
        streakAtRisk: { type: boolean, description: "true if streak will break tomorrow without study" }
        lastSevenDays:
          type: array
          items: { type: boolean }
          description: "7 booleans, oldest to newest (index 0 = 6 days ago, index 6 = today)"
...
    EarnedAchievementDto:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        icon: { type: string }
...